Added camera.launch.py
Added a launch for the camera node, which opens both the camera
node and the ros2 image previewer. Should've done this earlier,
as it makes a lot easier. It was necessary to add a launch to
change the video device parameter, as it was read-only.

camera.yaml
* added video_device parameter

camera.launch.py
* launch file to run v4l2 and rqt_image_viewer